+++ CHANGELOG.md
@@ -12,6 +12,12 @@ To install release candidates run `[sudo] gem install cocoapods --pre`
   [Samuel Giddins](https://github.com/segiddins)
   [#2390](https://github.com/CocoaPods/CocoaPods/issues/2390)
 
+* The `Installer` is passed directly to the `pre_install` and `post_install`
+  hooks defined in the Podfile, instead of the previously used
+  `Hooks::InstallerRepresentation`.  
+  [Marius Rackwitz](https://github.com/mrackwitz)
+  [#3648](https://github.com/CocoaPods/CocoaPods/issues/3648)
+
 ##### Enhancements
 
 * New commands `pod cache list` and `pod cache clean` allows you to see the